use local scripts to bump rc (remove lerna)
