--- lerna.json
@@ -1,22 +1,0 @@
-{
-  "version": "3.0.0-rc.11",
-  "npmClient": "pnpm",
-  "useWorkspaces": true,
-  "conventionalCommits": true,
-  "packages": ["packages/*"],
-  "command": {
-    "publish": {
-      "npmClient": "npm"
-    }
-  },
-  "changelog": {
-    "labels": {
-      "feat": "New Feature",
-      "fix": "Bug Fix",
-      "docs": "Documentation",
-      "types": "Types",
-      "perf": "Performance",
-      "refactor": "Refactor"
-    }
-  }
-}
