--- scripts/bump-rc.sh
@@ -1,12 +1,0 @@
-#!/bin/bash
-
-set -e
-
-# Restore all git changes
-git restore -s@ -SW  -- packages examples
-
-# Bump versions
-pnpm lerna version --preid rc --no-changelog --exact --no-git-tag-version --no-push -m "chore: release rc"
-
-# Stage
-git add --all
